Fix SLA computation
Invalid values are ignored and the SLA works modularly. For instance
if a /56 is provided, only 255 SLAs are possible if one assigns SLA
256 it loops to 0. If a /64 is provided, no SLAs are possible and the
id is ignored. A delegated /64 is only valid for a single interface.
Additional checks during the commit process need to be implemented to
ensure this cannot happen.